check file presence before copy

+++ scripts/prepare-test-containers.sh
@@ -5,5 +5,7 @@ BRANCH=${SB_BRANCH:-master}
 
 [ -d ${REPODIR} ] || git clone https://github.com/Juniper/contrail-test -b ${BRANCH}  ${REPODIR}
 cp common.env ${REPODIR}
-cp tpc.repo.template ${REPODIR}/docker/base/tpc.repo
-cp tpc.repo.template ${REPODIR}/docker/test/tpc.repo
+if [ -f tpc.repo.template ]; then
+  cp tpc.repo.template ${REPODIR}/docker/base/tpc.repo
+  cp tpc.repo.template ${REPODIR}/docker/test/tpc.repo
+fi
